First pass omnibot engine and game code.
Most gamecode is still ET based, I have not merged in the old game interface for JKA as I had to go digging in old caches for it.

Doesn't compile yet.

No Linux integration on engine yet for rendering.

Omni-bot repo must exist in the same root folder as this repository for time being. And must be named "omni-bot". Use develop branch.